Added main loop. Refactored code
Created a main device loop for all other sensors to be part of. Refactored camera and temperature code to have better file names, removed loop code within sensor files (only use main device loop from now on).
