+++ temperatureSensor.py
@@ -5,7 +5,7 @@ def sensor():
     for i in os.listdir('/sys/bus/w1/devices'):
         if i != 'w1_bus_master1':
             ds18b20 = i
-    return ds18b20
+            read(ds18b20)
 
 def read(ds18b20):
     location = '/sys/bus/w1/devices/' + ds18b20 + '/w1_slave'
@@ -22,13 +22,6 @@ def read(ds18b20):
 def loop(ds18b20):
     while True:
         if read(ds18b20) != None:
-            print "Current temperature : %0.2f C" % read(ds18b20)[0]
-            #print "Current temperature : %0.2f F" % read(ds18b20)[1]
+            #print "Current temperature : <temp here>"" %0.2f C" % read(ds18b20)[0]
+            print(f"Get temp here")
 
-
-if __name__ == '__main__':
-    try:
-        serialNum = sensor()
-        loop(serialNum)
-    except KeyboardInterrupt:
-        quit()
